pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hand_eye_calibration"
version = "0.0.1"
authors = [
    {name = "Ardeshir Talaei", email = "ardeshir.talaei@gmail.com"},
]
description = "A simple package for calculating the hand-eye calibration using OpenCV."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "BSD-3-Clause"}
dependencies = [
    "opencv-contrib-python",
    "numpy",
    "scipy",
    "pyyaml",
]

[project.optional-dependencies]
gui = ["eel",]

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["ui"]

[project.scripts]
HECalibration = "ui.__main__:main"
